Introduced VDOM.
We have added a layer between our code and the DOM. For that we defined the data structure of VDOM, made our "JSX" to create VDOM elements. And refactored the createElement to work with the VDOM definiton.
And since text is special - added some special handling of it to suit our needs.