And there we have it. A way of rerendering only the subtree that changed.
For this to work with the internal state - we had to give the Component a reference to his own VInstance. Kinda ugly but needed for maximum success!
What could be done next.. make it rerender only child components that actully need rerendeing in the changed subtree, implement lifecycle methods (or at least shouldComponentUpdate),  an actually beMuchSmarter method for reconcaliation. Well ... everything.